Convert "Deterministic builds with clang and lld" to Markdown
Fixes #52

Sort of. Articles pre-sometime in 2020 are in HTML form and with
https://github.com/JohannesKaufmann/html-to-markdown they convert
pretty well but not enough to mass convert them.

But since I put the effort into trying it I figured I would clean
up this one that had been noticed.

I used a script to reformat the header, html2markdown to
convert and then added back some plain text/monospace/code markers
where they had been missed, and removed some escapes that got added
because those plaintext bits weren't marked properly.

Now that it's Markdown, Hugo's theme will apply and produce the
right colours for light and dark mode.

The URL of the post will be the same.