Upgrade to newer fork of hugo-kiera theme
This fixes many issues when trying to build the site with the latest
Hugo. I tested this with Hugo v0.119.0 and latest at time of writing,
v0.139.0. It works with both.

https://github.com/avianto/hugo-kiera is unmaintainted and
https://themes.gohugo.io/themes/hugo-kiera/ lists https://github.com/funkydan2/hugo-kiera
as the active fork of it.

So I have imported that at commit
https://github.com/funkydan2/hugo-kiera/commit/5676dfaf565148f472d6a35c1f244d1199454a49.

I have also moved some things in our toml file:
* pagination has a new name
* author now needs to be in params.author

I have made no changes to anything in hugo-kiera/.